fix deleted images

+++ images/index.js
@@ -1,11 +1,14 @@
 const express = require("express");
 const fs = require('fs-extra');
-const bp = require('body-parser');
 const delay = require('delay');
 const app = express();
 
-app.use(express.json({limit: '50mb'}));
-app.use(express.urlencoded({limit: '50mb'}));
+app.use(express.json({
+  limit: '50mb'
+}));
+app.use(express.urlencoded({
+  limit: '50mb'
+}));
 
 app.post("/post", (req, res) => {
   const id = req.body.id
@@ -19,14 +22,18 @@ app.post("/post", (req, res) => {
 app.get("/gambar.jpg/:id/:i", (req, res) => {
   const id = req.params.id
   const i = req.params.i
-  res.sendFile(__dirname + `/gambar${id}${i}.jpg`, async () => {
-    await delay(300000)
-    try {
-    fs.unlinkSync(__dirname + `/gambar${id}${i}.jpg`)
-    } catch(e) {
-      
-    }
-  })
+  if (fs.existsSync(`gambar${id}${i}.jpg`)) {
+    res.sendFile(__dirname + `/gambar${id}${i}.jpg`, async () => {
+      await delay(300000)
+      try {
+        fs.unlinkSync(__dirname + `/gambar${id}${i}.jpg`)
+      } catch (e) {
+
+      }
+    })
+  } else {
+    res.send("Gambar Sudah Terhapus dari Server")
+  }
 })
 
 app.listen(process.env.PORT)
